+++ engine-io-snap/shell.nix
@@ -1,11 +1,31 @@
+{ nixpkgs ? import <nixpkgs> {}, compiler ? "default" }:
+
 let
-  pkgs = import <nixpkgs> {};
 
-  haskellPackages = pkgs.haskell-ng.packages.ghc7101.override {
-    overrides = self: super: {
-      engine-io = self.callPackage ../engine-io {};
-      engineIoSnap = self.callPackage ./. {};
-    };
-  };
+  inherit (nixpkgs) pkgs;
 
-in haskellPackages.engineIoSnap.env
+  f = { mkDerivation, base, bytestring, containers, engine-io
+      , io-streams, lifted-base, snap-core, stdenv, unordered-containers
+      , websockets, websockets-snap
+      }:
+      mkDerivation {
+        pname = "engine-io-snap";
+        version = "1.0.3";
+        src = ./.;
+        libraryHaskellDepends = [
+          base bytestring containers engine-io io-streams lifted-base
+          snap-core unordered-containers websockets websockets-snap
+        ];
+        homepage = "http://github.com/ocharles/engine.io";
+        license = stdenv.lib.licenses.bsd3;
+      };
+
+  haskellPackages = if compiler == "default"
+                       then pkgs.haskellPackages
+                       else pkgs.haskell.packages.${compiler};
+
+  drv = haskellPackages.callPackage f {};
+
+in
+
+  if pkgs.lib.inNixShell then drv.env else drv